copying to the right folder
